+++ src/client/rewards/rewardsActions.js
@@ -83,3 +83,18 @@ export const getBlacklist = userName => dispatch =>
     type: GET_BLACKLIST,
     payload: ApiClient.getBlacklist(userName),
   });
+
+export const GET_REWARDS_GENERAL_COUNTS = createAsyncActionType(
+  '@rewards/GET_REWARDS_GENERAL_COUNTS',
+);
+
+export const getRewardsGeneralCounts = () => (dispatch, getState) => {
+  const state = getState();
+  const authUserName = getAuthenticatedUserName(state);
+  return dispatch({
+    type: GET_REWARDS_GENERAL_COUNTS.ACTION,
+    payload: {
+      promise: ApiClient.getRewardsGeneralCounts({ userName: authUserName }),
+    },
+  });
+};
