+++ src/client/rewards/rewardsReducer.js
@@ -1,8 +1,16 @@
-import { SET_DATA_FOR_GLOBAL_REPORT, SET_DATA_FOR_SINGLE_REPORT } from './rewardsActions';
+import {
+  SET_DATA_FOR_GLOBAL_REPORT,
+  SET_DATA_FOR_SINGLE_REPORT,
+  GET_REWARDS_GENERAL_COUNTS,
+} from './rewardsActions';
 
 const initialState = {
   singleReportData: {},
   globalReportData: {},
+  tabType: '',
+  hasReceivables: '',
+  countTookPartCampaigns: 0,
+  createdCampaignsCount: 0,
 };
 
 const rewardsReducer = (state = initialState, action) => {
@@ -17,6 +25,15 @@ const rewardsReducer = (state = initialState, action) => {
         ...state,
         globalReportData: action.payload,
       };
+    case GET_REWARDS_GENERAL_COUNTS.SUCCESS: {
+      return {
+        ...state,
+        tabType: action.payload.tabType,
+        hasReceivables: action.payload.has_receivable,
+        countTookPartCampaigns: action.payload.count_took_part_campaigns,
+        createdCampaignsCount: action.payload.count_campaigns,
+      };
+    }
     default:
       return state;
   }
@@ -27,3 +44,7 @@ export default rewardsReducer;
 // export const getIsMapModalOpen = state => state.isFullscreenMode;
 export const getSingleReportData = state => state.singleReportData;
 export const getGlobalReportData = state => state.globalReportData;
+export const getTabType = state => state.tabType;
+export const getHasReceivables = state => state.hasReceivables;
+export const getCountTookPartCampaigns = state => state.countTookPartCampaigns;
+export const getCreatedCampaignsCount = state => state.createdCampaignsCount;
