+++ src/client/user/usersActions.js
@@ -142,18 +142,3 @@ export const updateUserMetadata = metadata => dispatch =>
     type: UPDATE_USER_METADATA,
     payload: metadata,
   });
-
-export const GET_REWARDS_GENERAL_COUNTS = createAsyncActionType(
-  '@users/GET_REWARDS_GENERAL_COUNTS',
-);
-
-export const getRewardsGeneralCounts = () => (dispatch, getState) => {
-  const state = getState();
-  const authUserName = getAuthenticatedUserName(state);
-  return dispatch({
-    type: GET_REWARDS_GENERAL_COUNTS.ACTION,
-    payload: {
-      promise: ApiClient.getRewardsGeneralCounts({ userName: authUserName }),
-    },
-  });
-};
